backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
manim==0.18.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
pinecone>=8.0.0
supabase>=2.3.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.5.3
aiofiles>=23.2.1
httpx>=0.24.0
